cmd/glue/py/requirements.txt
docopt==0.6.2
google==2.0.3
grpcio-tools==1.27.2
